+++ src/components/bookNote/preNote/PreNote.tsx
@@ -6,7 +6,8 @@ import styled, { css } from "styled-components";
 import { PeriNoteData, PreNoteData } from "../../../pages/BookNote";
 import { isLoginState } from "../../../utils/atom";
 import { PeriNoteTreeNode } from "../../../utils/dataType";
-import { patchBookNote, useFetchNote } from "../../../utils/useHooks";
+import { patchBookNote } from "../../../utils/lib/api";
+import { useFetchNote } from "../../../utils/useHooks";
 import { Loading } from "../../common";
 import { Button } from "../../common/styled/Button";
 import { PopUpPreDone, PreNoteForm, QuestionThree } from "..";
